+++ CHANGELOG.md
@@ -36,6 +36,17 @@
   * supports multiple assertions (nullability, greater, smaller, ...)
   * available under `dev.zieger.utils.core-testing.assertion2` (former `assert` utils remain in the library)
 
+##### 2.2.23
+
+* build fix
+
+##### 2.2.22
+
+* Continuation:
+    * allow multiple `suspendUntilTrigger` calls for the same trigger event
+    * made coroutine property nullable
+    * never launch a coroutine for unsuspended trigger call when no coroutine was defined
+
 ##### 2.2.21
 
 * build tools fix